Update python_threading.py
test2:
This example has a bit more code to explain, but it isn't that much more complicated than the first threading 

example, thanks to the use of the queuing module. This pattern is a very common and recommended way to use 

threads with Python. The steps are described as follows:

Create an instance of Queue.Queue() and then populate it with data.
Pass that instance of populated data into the threading class that you created from inheriting from 

threading.Thread.
Spawn a pool of daemon threads.
Pull one item out of the queue at a time, and use that data inside of the thread, the run method, to do the 

work.
After the work is done, send a signal to the queue with queue.task_done() that the task has been completed.
Join on the queue, which really means to wait until the queue is empty, and then exit the main program.
Just a note about this pattern: By setting daemonic threads to true, it allows the main thread, or program, 

to exit if only daemonic threads are alive. This creates a simple way to control the flow of the program, 

because you can then join on the queue, or wait until the queue is empty, before exiting. The exact process 

is best described in the documentation for the queue module, as seen in the Related topics:

join()
"Blocks until all items in the queue have been gotten and processed. The count of unfinished tasks goes up 

whenever an item is added to the queue. The count goes down whenever a consumer thread calls task_done() to 

indicate that the item was retrieved and all work on it is complete. When the count of unfinished tasks drops 

to zero, join() unblocks.